Config: Blue short: Next predefined settings Blue long: Load default values Added predefined columns to the config struct: Performance, Information
To define an additional predefined setting, just add a new row to the predefined settings dict and the corresponding config column.

Notes on the config column: A config element with a none value will be ignored and never changed.

Load defaults will reload all default values.

The blue button will indicate the next predefined setting, which can be loaded.

The title gets a <ID> marker, if the config matches a predefined setting.

Untested

git-svn-id: svn://85.10.198.214/trunk@419 398de0b3-2755-4f39-ad97-b6641548fef7